requirements.txt
langchain>=0.2.10
langchain-openai>=0.3.0
python-dotenv>=1.0.0